Storage Package Tests: * Create test/changelist_data/storage/conftest.py * Create test/changelist_data/storage/test_method_load_storage.py * Create test/changelist_data/storage/test_method_read_storage.py * Update test/changelist_data/storage/test_changelist_data_storage.py - exported fixtures to conftest * Update test/changelist_data/storage/test_changelists_storage.py - exported fixtures to conftest * Update test/changelist_data/storage/test_file_validation.py - replace provider methods with fixtures from conftest * Remove test/changelist_data/storage/test_init.py - split into test_method_ files for each package level method * Update test/changelist_data/storage/test_workspace_storage.py - exported fixtures to conftest, replace provider methods with fixtures from conftest
