+++ test/changelist_data/storage/conftest.py
@@ -1,0 +1,19 @@
+"""
+"""
+import tempfile
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture
+def temp_dir():
+    tdir = tempfile.TemporaryDirectory()
+    yield tdir
+    tdir.cleanup()
+
+
+@pytest.fixture
+def temp_file(temp_dir):
+    file = temp_dir.name + "/test_xml.xml"
+    yield Path(file)
